requirements.txt
numpy
Pillow
scikit-image
opencv-python
streamlit
matplotlib
streamlit_image_select
pandas
openpyxl